project/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
asyncio==4.0.0
certifi==2025.8.3
colorama==0.4.6
distro==1.9.0
Django==5.2.5
djangorestframework==3.16.1
drf-yasg==1.21.10
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
inflection==0.5.1
jiter==0.10.0
openai==1.100.2
packaging==25.0
pillow==11.3.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
gunicorn==22.0.0